kelar juga yey tapi crud cuma di barang, sisanya secukupnya aja

+++ app/Config/Routes.php
@@ -35,9 +35,17 @@ $routes->get('/', 'Home::index');
 $routes->get('/login', 'Login::index');
 $routes->get('/logout', 'Login::logout');
 // dashboard
-$routes->get('/dashboard', 'Dashboard::index');
+$routes->get('/dashboard', 'Dashboard::index', ['filter' => 'auth']);
 // barang
-$routes->delete('barang/delete/(:any)', 'Barang::delete/$1');
+$routes->get('barang', 'Barang::index', ['filter' => 'auth']);
+$routes->delete('barang/delete/(:any)', 'Barang::delete/$1', ['filter' => 'auth']);
+$routes->get('barang/edit/(:any)', 'Barang::edit/$1', ['filter' => 'auth']);
+$routes->get('barang/detail/(:any)', 'Barang::detail/$1', ['filter' => 'auth']);
+//pinjam
+$routes->get('peminjaman', 'Peminjaman::index', ['filter' => 'auth']);
+$routes->get('peminjaman/detail/(:any)', 'Peminjaman::detail/$1', ['filter' => 'auth']);
+//lokasi
+$routes->get('lokasi', 'Lokasi::index', ['filter' => 'auth']);
 
 /*
  * --------------------------------------------------------------------
